+++ devices/hip/hip_device.cpp
@@ -105,8 +105,8 @@ OIDN_NAMESPACE_BEGIN
     : deviceID(deviceID),
       stream(stream)
   {
-    if (deviceID < 0)
-      checkError(hipGetDevice(&this->deviceID));
+    if (deviceID < 0) // deprecated behavior
+      this->deviceID = 0;
   }
 
   HIPDevice::HIPDevice(const Ref<HIPPhysicalDevice>& physicalDevice)